debuging some stuff: turns out - the mega has some problems with interrupts on some ports. i am unable to get the interrupts working on another port than D2
so as a result of this, i have to change the code so that RPMs are not a thing anymore
the speed is therefore only adjusted by using PWM in a defined range without considering RPM as a second input.

in the end - this makes thing easier by maintaing the same function.
so, 3h well spend :/